move wildcard expanding tests to expand_wildcard_rule.rs
